refactor(gateway): split local demo seed fixtures
