+++ crates/gateway/src/local_demo_seed/api_keys.rs
@@ -1,0 +1,123 @@
+use super::{LocalDemoApiKeyFixture, LocalDemoOwnerFixture};
+
+pub(super) const LOCAL_DEMO_API_KEYS: &[LocalDemoApiKeyFixture] = &[
+    LocalDemoApiKeyFixture {
+        name: "Alice Personal Key",
+        public_id: "locdemoalice1",
+        secret: "alice-demo-secret",
+        owner: LocalDemoOwnerFixture::User("alice@platform.local"),
+        model_keys: &["openai-fast", "openai-fast-v2"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Cara Research Key",
+        public_id: "locdemocara1",
+        secret: "cara-demo-secret",
+        owner: LocalDemoOwnerFixture::User("cara@research.local"),
+        model_keys: &["gemini-fast", "claude-sonnet"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Diego Daily Budget Key",
+        public_id: "locdemodiego1",
+        secret: "diego-demo-secret",
+        owner: LocalDemoOwnerFixture::User("diego@research.local"),
+        model_keys: &["openai-fast", "claude-sonnet"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Platform Team Key",
+        public_id: "locdemoplatform1",
+        secret: "platform-demo-secret",
+        owner: LocalDemoOwnerFixture::Team("platform"),
+        model_keys: &["openai-fast-v2", "gemini-fast"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Ben Analyst Key",
+        public_id: "locdemoben1",
+        secret: "ben-demo-secret",
+        owner: LocalDemoOwnerFixture::User("ben@platform.local"),
+        model_keys: &["openai-fast-v2", "gemini-pro-preview"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Erin Evaluation Key",
+        public_id: "locdemoerin1",
+        secret: "erin-demo-secret",
+        owner: LocalDemoOwnerFixture::User("erin@research.local"),
+        model_keys: &["claude-opus", "gemini-fast"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Faye Research Key",
+        public_id: "locdemofaye1",
+        secret: "faye-demo-secret",
+        owner: LocalDemoOwnerFixture::User("faye@research.local"),
+        model_keys: &["claude-opus", "bedrock-claude-sonnet"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Gina Tooling Key",
+        public_id: "locdemogina1",
+        secret: "gina-demo-secret",
+        owner: LocalDemoOwnerFixture::User("gina@research.local"),
+        model_keys: &["gpt-oss-120b", "gemini-pro-preview"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Hugo Analysis Key",
+        public_id: "locdemohugo1",
+        secret: "hugo-demo-secret",
+        owner: LocalDemoOwnerFixture::User("hugo@research.local"),
+        model_keys: &["claude-sonnet", "gpt-oss-120b"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Ivy Coordination Key",
+        public_id: "locdemoivy1",
+        secret: "ivy-demo-secret",
+        owner: LocalDemoOwnerFixture::User("ivy@research.local"),
+        model_keys: &["gemini-fast", "gemini-pro-preview"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Jules Operations Research Key",
+        public_id: "locdemojules1",
+        secret: "jules-demo-secret",
+        owner: LocalDemoOwnerFixture::User("jules@research.local"),
+        model_keys: &["claude-sonnet", "bedrock-claude-sonnet"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Frank Applied Key",
+        public_id: "locdemofrank1",
+        secret: "frank-demo-secret",
+        owner: LocalDemoOwnerFixture::User("frank@applied.local"),
+        model_keys: &["gpt-oss-120b", "openai-fast-v2"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Grace Applied Key",
+        public_id: "locdemograce1",
+        secret: "grace-demo-secret",
+        owner: LocalDemoOwnerFixture::User("grace@applied.local"),
+        model_keys: &["bedrock-claude-sonnet", "gemini-pro-preview"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Heidi Applied Key",
+        public_id: "locdemoheidi1",
+        secret: "heidi-demo-secret",
+        owner: LocalDemoOwnerFixture::User("heidi@applied.local"),
+        model_keys: &["openai-fast", "claude-opus"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Ivan Operations Key",
+        public_id: "locdemoivan1",
+        secret: "ivan-demo-secret",
+        owner: LocalDemoOwnerFixture::User("ivan@operations.local"),
+        model_keys: &["gemini-fast", "gpt-oss-120b"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Judy Operations Key",
+        public_id: "locdemojudy1",
+        secret: "judy-demo-secret",
+        owner: LocalDemoOwnerFixture::User("judy@operations.local"),
+        model_keys: &["bedrock-claude-sonnet", "openai-fast-v2"],
+    },
+    LocalDemoApiKeyFixture {
+        name: "Mallory Operations Key",
+        public_id: "locdemomallory1",
+        secret: "mallory-demo-secret",
+        owner: LocalDemoOwnerFixture::User("mallory@operations.local"),
+        model_keys: &["openai-fast-v2", "gemini-pro-preview"],
+    },
+];
