+++ crates/gateway/src/local_demo_seed/models.rs
@@ -1,0 +1,10 @@
+pub(super) const LOCAL_DEMO_MODEL_KEYS: &[&str] = &[
+    "openai-fast",
+    "openai-fast-v2",
+    "gemini-fast",
+    "claude-sonnet",
+    "claude-opus",
+    "bedrock-claude-sonnet",
+    "gpt-oss-120b",
+    "gemini-pro-preview",
+];
