+++ crates/gateway/src/local_demo_seed/teams.rs
@@ -1,0 +1,2 @@
+pub(super) const LOCAL_DEMO_TEAM_KEYS: &[&str] =
+    &["platform", "research", "applied-ai", "operations"];
