+++ crates/gateway/src/local_demo_seed/usage.rs
@@ -1,0 +1,928 @@
+use super::LocalDemoRequestFixture;
+
+pub(super) const LOCAL_DEMO_REQUESTS: &[LocalDemoRequestFixture] = &[
+    LocalDemoRequestFixture {
+        request_id: "demo-req-001",
+        api_key_public_id: "locdemoalice1",
+        days_ago: 0,
+        hours_ago: 2,
+        model_key: "openai-fast",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-prod",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(620),
+        completion_tokens: Some(280),
+        cost_scaled: 1_245,
+        status_code: 200,
+        latency_ms: 482,
+        service: "admin-ui",
+        component: "dashboard",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "summary",
+        prompt: "Summarize this week’s API key usage.",
+        completion: "Here is a compact weekly usage summary for the platform team.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-002",
+        api_key_public_id: "locdemoplatform1",
+        days_ago: 0,
+        hours_ago: 4,
+        model_key: "openai-fast-v2",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-secondary",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(840),
+        completion_tokens: Some(410),
+        cost_scaled: 2_180,
+        status_code: 200,
+        latency_ms: 611,
+        service: "batch-jobs",
+        component: "nightly-rollup",
+        env: "local",
+        bespoke_key: "job",
+        bespoke_value: "cost-rollup",
+        prompt: "Generate a nightly rollup of usage and spend anomalies.",
+        completion: "Nightly rollup completed with two notable spend anomalies.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-003",
+        api_key_public_id: "locdemocara1",
+        days_ago: 1,
+        hours_ago: 3,
+        model_key: "gemini-fast",
+        resolved_model_key: "gemini-fast",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-2.0-flash",
+        prompt_tokens: Some(510),
+        completion_tokens: Some(320),
+        cost_scaled: 1_610,
+        status_code: 200,
+        latency_ms: 729,
+        service: "research-lab",
+        component: "experiment-runner",
+        env: "local",
+        bespoke_key: "experiment",
+        bespoke_value: "safety-eval",
+        prompt: "Draft notes for the latest safety evaluation batch.",
+        completion: "The latest safety evaluation batch shows stable refusal behavior.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-004",
+        api_key_public_id: "locdemodiego1",
+        days_ago: 0,
+        hours_ago: 1,
+        model_key: "claude-sonnet",
+        resolved_model_key: "claude-sonnet",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-sonnet-4-6",
+        prompt_tokens: Some(430),
+        completion_tokens: Some(690),
+        cost_scaled: 3_420,
+        status_code: 200,
+        latency_ms: 938,
+        service: "research-lab",
+        component: "daily-brief",
+        env: "local",
+        bespoke_key: "brief",
+        bespoke_value: "morning",
+        prompt: "Write the daily research brief for active experiments.",
+        completion: "Daily research brief prepared for the active experiment queue.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-005",
+        api_key_public_id: "locdemoalice1",
+        days_ago: 2,
+        hours_ago: 6,
+        model_key: "openai-fast-v2",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-prod",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(710),
+        completion_tokens: Some(360),
+        cost_scaled: 1_980,
+        status_code: 200,
+        latency_ms: 564,
+        service: "admin-ui",
+        component: "request-log-detail",
+        env: "local",
+        bespoke_key: "panel",
+        bespoke_value: "detail",
+        prompt: "Explain the request log payload delta between retries.",
+        completion: "The second attempt used the fallback provider and a shorter system prompt.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-006",
+        api_key_public_id: "locdemoplatform1",
+        days_ago: 3,
+        hours_ago: 5,
+        model_key: "gemini-fast",
+        resolved_model_key: "gemini-fast",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-2.0-flash",
+        prompt_tokens: Some(390),
+        completion_tokens: Some(240),
+        cost_scaled: 1_140,
+        status_code: 200,
+        latency_ms: 688,
+        service: "batch-jobs",
+        component: "owner-sync",
+        env: "local",
+        bespoke_key: "job",
+        bespoke_value: "owner-sync",
+        prompt: "Normalize owner metadata for request-log exports.",
+        completion: "Owner metadata normalized for the latest export window.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-007",
+        api_key_public_id: "locdemocara1",
+        days_ago: 4,
+        hours_ago: 2,
+        model_key: "claude-sonnet",
+        resolved_model_key: "claude-sonnet",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-sonnet-4-6",
+        prompt_tokens: Some(560),
+        completion_tokens: Some(470),
+        cost_scaled: 2_860,
+        status_code: 200,
+        latency_ms: 1_024,
+        service: "research-lab",
+        component: "retrospective",
+        env: "local",
+        bespoke_key: "artifact",
+        bespoke_value: "retrospective",
+        prompt: "Write a concise retrospective for the last experiment iteration.",
+        completion: "The retrospective highlights routing stability and lower-than-expected cost.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-008",
+        api_key_public_id: "locdemodiego1",
+        days_ago: 1,
+        hours_ago: 7,
+        model_key: "openai-fast",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-prod",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(280),
+        completion_tokens: Some(150),
+        cost_scaled: 790,
+        status_code: 200,
+        latency_ms: 431,
+        service: "research-lab",
+        component: "token-audit",
+        env: "local",
+        bespoke_key: "audit",
+        bespoke_value: "token-usage",
+        prompt: "Audit the latest token usage sample for anomalies.",
+        completion: "Token usage stayed within the expected variance band.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-009",
+        api_key_public_id: "locdemoplatform1",
+        days_ago: 5,
+        hours_ago: 4,
+        model_key: "openai-fast-v2",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-prod",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(930),
+        completion_tokens: Some(510),
+        cost_scaled: 2_640,
+        status_code: 200,
+        latency_ms: 702,
+        service: "batch-jobs",
+        component: "forecasting",
+        env: "local",
+        bespoke_key: "job",
+        bespoke_value: "forecasting",
+        prompt: "Forecast next week’s spend trend from the last 14 days.",
+        completion: "The next week trends slightly upward with stable model mix.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-010",
+        api_key_public_id: "locdemoalice1",
+        days_ago: 6,
+        hours_ago: 3,
+        model_key: "openai-fast",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-secondary",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(470),
+        completion_tokens: Some(250),
+        cost_scaled: 1_120,
+        status_code: 200,
+        latency_ms: 547,
+        service: "admin-ui",
+        component: "breadcrumbs",
+        env: "local",
+        bespoke_key: "screen",
+        bespoke_value: "overview",
+        prompt: "Summarize top navigation usage patterns from the admin UI.",
+        completion: "The overview screen remains the most active admin entry point.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-011",
+        api_key_public_id: "locdemocara1",
+        days_ago: 2,
+        hours_ago: 8,
+        model_key: "gemini-fast",
+        resolved_model_key: "gemini-fast",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-2.0-flash",
+        prompt_tokens: None,
+        completion_tokens: None,
+        cost_scaled: 0,
+        status_code: 429,
+        latency_ms: 214,
+        service: "research-lab",
+        component: "burst-eval",
+        env: "local",
+        bespoke_key: "failure",
+        bespoke_value: "rate-limit",
+        prompt: "Run the burst evaluation batch.",
+        completion: "rate limit exceeded by upstream",
+        error_code: Some("rate_limited"),
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-012",
+        api_key_public_id: "locdemodiego1",
+        days_ago: 3,
+        hours_ago: 9,
+        model_key: "claude-sonnet",
+        resolved_model_key: "claude-sonnet",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-sonnet-4-6",
+        prompt_tokens: None,
+        completion_tokens: None,
+        cost_scaled: 0,
+        status_code: 502,
+        latency_ms: 301,
+        service: "research-lab",
+        component: "provider-fallback",
+        env: "local",
+        bespoke_key: "failure",
+        bespoke_value: "upstream-error",
+        prompt: "Generate a fallback plan for the next routing attempt.",
+        completion: "upstream gateway timeout",
+        error_code: Some("upstream_http_502"),
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-013",
+        api_key_public_id: "locdemoben1",
+        days_ago: 0,
+        hours_ago: 5,
+        model_key: "gemini-pro-preview",
+        resolved_model_key: "gemini-pro-preview",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-3.1-pro-preview",
+        prompt_tokens: Some(1_120),
+        completion_tokens: Some(530),
+        cost_scaled: 2_940,
+        status_code: 200,
+        latency_ms: 812,
+        service: "admin-ui",
+        component: "model-inventory",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "inventory-review",
+        prompt: "Compare routed model inventory against the current provider catalog.",
+        completion: "The routed inventory now includes Vertex preview coverage and Bedrock entries.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-014",
+        api_key_public_id: "locdemoerin1",
+        days_ago: 1,
+        hours_ago: 2,
+        model_key: "claude-opus",
+        resolved_model_key: "claude-opus",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-opus-4-7",
+        prompt_tokens: Some(1_880),
+        completion_tokens: Some(940),
+        cost_scaled: 7_650,
+        status_code: 200,
+        latency_ms: 1_348,
+        service: "research-lab",
+        component: "longform-eval",
+        env: "local",
+        bespoke_key: "experiment",
+        bespoke_value: "reasoning-depth",
+        prompt: "Review the longform reasoning transcript and flag evaluation gaps.",
+        completion: "The transcript shows two evaluation gaps around tool grounding and citations.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-015",
+        api_key_public_id: "locdemofaye1",
+        days_ago: 2,
+        hours_ago: 1,
+        model_key: "bedrock-claude-sonnet",
+        resolved_model_key: "bedrock-claude-sonnet",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "claude-sonnet-4-6",
+        prompt_tokens: Some(760),
+        completion_tokens: Some(450),
+        cost_scaled: 3_380,
+        status_code: 200,
+        latency_ms: 1_102,
+        service: "research-lab",
+        component: "bedrock-routing",
+        env: "local",
+        bespoke_key: "experiment",
+        bespoke_value: "bedrock-claude",
+        prompt: "Summarize Bedrock Claude route behavior for the latest eval batch.",
+        completion: "Bedrock Claude completed the batch with stable latency and moderate cost.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-016",
+        api_key_public_id: "locdemogina1",
+        days_ago: 0,
+        hours_ago: 8,
+        model_key: "gpt-oss-120b",
+        resolved_model_key: "gpt-oss-120b",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "gpt-oss-120b",
+        prompt_tokens: Some(1_340),
+        completion_tokens: Some(620),
+        cost_scaled: 4_210,
+        status_code: 200,
+        latency_ms: 978,
+        service: "research-lab",
+        component: "tool-runner",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "mcp-analysis",
+        prompt: "Use the available MCP context to rank the latest experiment artifacts.",
+        completion: "The artifacts were ranked by reproducibility, coverage, and freshness.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-017",
+        api_key_public_id: "locdemohugo1",
+        days_ago: 3,
+        hours_ago: 4,
+        model_key: "gpt-oss-120b",
+        resolved_model_key: "gpt-oss-120b",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "gpt-oss-120b",
+        prompt_tokens: Some(910),
+        completion_tokens: Some(380),
+        cost_scaled: 2_760,
+        status_code: 200,
+        latency_ms: 724,
+        service: "research-lab",
+        component: "metric-synthesis",
+        env: "local",
+        bespoke_key: "artifact",
+        bespoke_value: "leaderboard",
+        prompt: "Condense user-level usage metrics into a leaderboard narrative.",
+        completion: "Usage is concentrated in applied and research workflows with broad MCP activity.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-018",
+        api_key_public_id: "locdemofrank1",
+        days_ago: 0,
+        hours_ago: 3,
+        model_key: "gpt-oss-120b",
+        resolved_model_key: "gpt-oss-120b",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "gpt-oss-120b",
+        prompt_tokens: Some(2_300),
+        completion_tokens: Some(1_050),
+        cost_scaled: 8_740,
+        status_code: 200,
+        latency_ms: 1_186,
+        service: "applied-ai",
+        component: "customer-brief",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "customer-brief",
+        prompt: "Draft a customer-facing brief from product telemetry and support notes.",
+        completion: "The brief emphasizes reduced latency, steadier routing, and lower error volume.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-019",
+        api_key_public_id: "locdemograce1",
+        days_ago: 1,
+        hours_ago: 6,
+        model_key: "gemini-pro-preview",
+        resolved_model_key: "gemini-pro-preview",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-3.1-pro-preview",
+        prompt_tokens: Some(680),
+        completion_tokens: Some(260),
+        cost_scaled: 1_890,
+        status_code: 200,
+        latency_ms: 603,
+        service: "applied-ai",
+        component: "prototype-review",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "prototype",
+        prompt: "Review prototype feedback and group the top product risks.",
+        completion: "Top risks cluster around onboarding clarity, routing explainability, and budgets.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-020",
+        api_key_public_id: "locdemoheidi1",
+        days_ago: 4,
+        hours_ago: 5,
+        model_key: "claude-opus",
+        resolved_model_key: "claude-opus",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-opus-4-7",
+        prompt_tokens: Some(1_560),
+        completion_tokens: Some(720),
+        cost_scaled: 6_130,
+        status_code: 200,
+        latency_ms: 1_261,
+        service: "applied-ai",
+        component: "design-critique",
+        env: "local",
+        bespoke_key: "screen",
+        bespoke_value: "budgets",
+        prompt: "Critique the current budget page information hierarchy.",
+        completion: "The page should separate team-level controls from personal spend signals.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-021",
+        api_key_public_id: "locdemoivan1",
+        days_ago: 0,
+        hours_ago: 9,
+        model_key: "gemini-fast",
+        resolved_model_key: "gemini-fast",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-2.0-flash",
+        prompt_tokens: Some(420),
+        completion_tokens: Some(190),
+        cost_scaled: 1_020,
+        status_code: 200,
+        latency_ms: 388,
+        service: "operations",
+        component: "incident-review",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "incident",
+        prompt: "Summarize the latest provider incident notes for the on-call handoff.",
+        completion: "The incident was isolated to a transient provider timeout window.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-022",
+        api_key_public_id: "locdemojudy1",
+        days_ago: 5,
+        hours_ago: 7,
+        model_key: "bedrock-claude-sonnet",
+        resolved_model_key: "bedrock-claude-sonnet",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "claude-sonnet-4-6",
+        prompt_tokens: Some(350),
+        completion_tokens: Some(130),
+        cost_scaled: 920,
+        status_code: 200,
+        latency_ms: 507,
+        service: "operations",
+        component: "runbook-update",
+        env: "local",
+        bespoke_key: "artifact",
+        bespoke_value: "runbook",
+        prompt: "Update the Bedrock fallback runbook with current route names.",
+        completion: "The fallback runbook now references Bedrock Claude and GPT OSS routes.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-023",
+        api_key_public_id: "locdemogina1",
+        days_ago: 2,
+        hours_ago: 10,
+        model_key: "gemini-pro-preview",
+        resolved_model_key: "gemini-pro-preview",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-3.1-pro-preview",
+        prompt_tokens: None,
+        completion_tokens: None,
+        cost_scaled: 0,
+        status_code: 503,
+        latency_ms: 342,
+        service: "research-lab",
+        component: "tool-runner",
+        env: "local",
+        bespoke_key: "failure",
+        bespoke_value: "provider-unavailable",
+        prompt: "Run the preview model comparison with MCP artifact lookup.",
+        completion: "preview provider temporarily unavailable",
+        error_code: Some("upstream_http_503"),
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-024",
+        api_key_public_id: "locdemofrank1",
+        days_ago: 2,
+        hours_ago: 2,
+        model_key: "gpt-oss-120b",
+        resolved_model_key: "gpt-oss-120b",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "gpt-oss-120b",
+        prompt_tokens: Some(1_740),
+        completion_tokens: Some(810),
+        cost_scaled: 6_490,
+        status_code: 200,
+        latency_ms: 1_054,
+        service: "applied-ai",
+        component: "customer-brief",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "renewal-risk",
+        prompt: "Summarize renewal risk from support notes and recent telemetry.",
+        completion: "Renewal risk is concentrated around onboarding delays and budget alerts.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-025",
+        api_key_public_id: "locdemoerin1",
+        days_ago: 2,
+        hours_ago: 4,
+        model_key: "claude-opus",
+        resolved_model_key: "claude-opus",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-opus-4-7",
+        prompt_tokens: Some(2_120),
+        completion_tokens: Some(1_180),
+        cost_scaled: 9_880,
+        status_code: 200,
+        latency_ms: 1_482,
+        service: "research-lab",
+        component: "longform-eval",
+        env: "local",
+        bespoke_key: "experiment",
+        bespoke_value: "eval-review",
+        prompt: "Review the evaluation traces and identify recurring failure modes.",
+        completion: "Recurring failures cluster around incomplete evidence and stale context.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-026",
+        api_key_public_id: "locdemoivy1",
+        days_ago: 2,
+        hours_ago: 7,
+        model_key: "gemini-pro-preview",
+        resolved_model_key: "gemini-pro-preview",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-3.1-pro-preview",
+        prompt_tokens: Some(640),
+        completion_tokens: Some(260),
+        cost_scaled: 1_770,
+        status_code: 200,
+        latency_ms: 522,
+        service: "research-lab",
+        component: "experiment-runner",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "queue-triage",
+        prompt: "Group pending experiments by owner and expected review effort.",
+        completion: "Experiments were grouped into quick checks, reruns, and deep reviews.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-027",
+        api_key_public_id: "locdemojules1",
+        days_ago: 3,
+        hours_ago: 1,
+        model_key: "bedrock-claude-sonnet",
+        resolved_model_key: "bedrock-claude-sonnet",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "claude-sonnet-4-6",
+        prompt_tokens: Some(940),
+        completion_tokens: Some(510),
+        cost_scaled: 3_720,
+        status_code: 200,
+        latency_ms: 887,
+        service: "research-lab",
+        component: "bedrock-routing",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "route-comparison",
+        prompt: "Compare Vertex and Bedrock Claude route outcomes for the eval suite.",
+        completion: "Bedrock had higher median latency while Vertex produced longer answers.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-028",
+        api_key_public_id: "locdemogina1",
+        days_ago: 3,
+        hours_ago: 3,
+        model_key: "gpt-oss-120b",
+        resolved_model_key: "gpt-oss-120b",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "gpt-oss-120b",
+        prompt_tokens: Some(1_010),
+        completion_tokens: Some(390),
+        cost_scaled: 2_980,
+        status_code: 200,
+        latency_ms: 742,
+        service: "research-lab",
+        component: "tool-runner",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "artifact-scoring",
+        prompt: "Score MCP-linked artifacts by freshness and completeness.",
+        completion: "Artifact scores prioritize complete traces with recent source updates.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-029",
+        api_key_public_id: "locdemomallory1",
+        days_ago: 3,
+        hours_ago: 6,
+        model_key: "openai-fast-v2",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-prod",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(520),
+        completion_tokens: Some(180),
+        cost_scaled: 1_430,
+        status_code: 200,
+        latency_ms: 468,
+        service: "operations",
+        component: "runbook-update",
+        env: "local",
+        bespoke_key: "artifact",
+        bespoke_value: "handoff",
+        prompt: "Prepare an on-call handoff from the latest request-log sample.",
+        completion: "The handoff highlights one provider incident and two budget warnings.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-030",
+        api_key_public_id: "locdemoalice1",
+        days_ago: 4,
+        hours_ago: 1,
+        model_key: "openai-fast",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-secondary",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(860),
+        completion_tokens: Some(330),
+        cost_scaled: 2_150,
+        status_code: 200,
+        latency_ms: 534,
+        service: "admin-ui",
+        component: "dashboard",
+        env: "local",
+        bespoke_key: "screen",
+        bespoke_value: "leaderboard",
+        prompt: "Summarize leaderboard movement over the past week.",
+        completion: "Applied AI moved above research after several high-context requests.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-031",
+        api_key_public_id: "locdemofaye1",
+        days_ago: 4,
+        hours_ago: 4,
+        model_key: "claude-opus",
+        resolved_model_key: "claude-opus",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-opus-4-7",
+        prompt_tokens: Some(1_450),
+        completion_tokens: Some(610),
+        cost_scaled: 5_570,
+        status_code: 200,
+        latency_ms: 1_214,
+        service: "research-lab",
+        component: "longform-eval",
+        env: "local",
+        bespoke_key: "experiment",
+        bespoke_value: "rubric-tuning",
+        prompt: "Tune the evaluation rubric using the previous analyst comments.",
+        completion: "The rubric now separates evidence quality from reasoning quality.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-032",
+        api_key_public_id: "locdemoben1",
+        days_ago: 4,
+        hours_ago: 8,
+        model_key: "gemini-pro-preview",
+        resolved_model_key: "gemini-pro-preview",
+        provider_key: "vertex-adc",
+        upstream_model: "google/gemini-3.1-pro-preview",
+        prompt_tokens: Some(390),
+        completion_tokens: Some(160),
+        cost_scaled: 980,
+        status_code: 200,
+        latency_ms: 376,
+        service: "admin-ui",
+        component: "model-inventory",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "model-audit",
+        prompt: "List models that should be reviewed before budget planning.",
+        completion: "Claude Opus and GPT OSS should be reviewed because of recent spend spikes.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-033",
+        api_key_public_id: "locdemoheidi1",
+        days_ago: 5,
+        hours_ago: 2,
+        model_key: "claude-opus",
+        resolved_model_key: "claude-opus",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-opus-4-7",
+        prompt_tokens: Some(1_190),
+        completion_tokens: Some(470),
+        cost_scaled: 4_610,
+        status_code: 200,
+        latency_ms: 1_038,
+        service: "applied-ai",
+        component: "design-critique",
+        env: "local",
+        bespoke_key: "screen",
+        bespoke_value: "teams",
+        prompt: "Identify which team page sections become crowded with more users.",
+        completion: "Member tables and budget ownership controls need stronger separation.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-034",
+        api_key_public_id: "locdemograce1",
+        days_ago: 5,
+        hours_ago: 5,
+        model_key: "bedrock-claude-sonnet",
+        resolved_model_key: "bedrock-claude-sonnet",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "claude-sonnet-4-6",
+        prompt_tokens: Some(730),
+        completion_tokens: Some(280),
+        cost_scaled: 2_240,
+        status_code: 200,
+        latency_ms: 646,
+        service: "applied-ai",
+        component: "prototype-review",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "scenario-review",
+        prompt: "Review prototype scenarios and flag missing admin states.",
+        completion: "Missing states include empty grants, expired keys, and dense teams.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-035",
+        api_key_public_id: "locdemoivan1",
+        days_ago: 5,
+        hours_ago: 9,
+        model_key: "gpt-oss-120b",
+        resolved_model_key: "gpt-oss-120b",
+        provider_key: "bedrock-us-east-1",
+        upstream_model: "gpt-oss-120b",
+        prompt_tokens: Some(610),
+        completion_tokens: Some(250),
+        cost_scaled: 1_860,
+        status_code: 200,
+        latency_ms: 558,
+        service: "operations",
+        component: "incident-review",
+        env: "local",
+        bespoke_key: "workflow",
+        bespoke_value: "postmortem",
+        prompt: "Draft the postmortem summary for the provider timeout window.",
+        completion: "The postmortem summary covers detection, fallback, and follow-up work.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-036",
+        api_key_public_id: "locdemocara1",
+        days_ago: 6,
+        hours_ago: 2,
+        model_key: "claude-sonnet",
+        resolved_model_key: "claude-sonnet",
+        provider_key: "vertex-claude",
+        upstream_model: "anthropic/claude-sonnet-4-6",
+        prompt_tokens: Some(880),
+        completion_tokens: Some(420),
+        cost_scaled: 3_120,
+        status_code: 200,
+        latency_ms: 932,
+        service: "research-lab",
+        component: "retrospective",
+        env: "local",
+        bespoke_key: "artifact",
+        bespoke_value: "weekly-retro",
+        prompt: "Summarize the weekly research retrospective for leadership.",
+        completion: "Research velocity improved while route diversity increased operating cost.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-037",
+        api_key_public_id: "locdemojudy1",
+        days_ago: 6,
+        hours_ago: 5,
+        model_key: "openai-fast-v2",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-secondary",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(460),
+        completion_tokens: Some(170),
+        cost_scaled: 1_210,
+        status_code: 200,
+        latency_ms: 421,
+        service: "operations",
+        component: "runbook-update",
+        env: "local",
+        bespoke_key: "artifact",
+        bespoke_value: "budget-playbook",
+        prompt: "Update the budget alert playbook using this week's examples.",
+        completion: "The playbook now includes examples for soft warnings and hard stops.",
+        error_code: None,
+    },
+    LocalDemoRequestFixture {
+        request_id: "demo-req-038",
+        api_key_public_id: "locdemodiego1",
+        days_ago: 6,
+        hours_ago: 7,
+        model_key: "openai-fast",
+        resolved_model_key: "openai-fast-v2",
+        provider_key: "openai-prod",
+        upstream_model: "gpt-5",
+        prompt_tokens: Some(330),
+        completion_tokens: Some(140),
+        cost_scaled: 880,
+        status_code: 200,
+        latency_ms: 392,
+        service: "research-lab",
+        component: "token-audit",
+        env: "local",
+        bespoke_key: "audit",
+        bespoke_value: "weekly-token-sample",
+        prompt: "Audit the weekly token sample and call out unusual spikes.",
+        completion: "Two spikes came from longform eval prompts and one from MCP artifact scoring.",
+        error_code: None,
+    },
+];
+
+pub(super) fn demo_tool_cardinality(
+    fixture: &LocalDemoRequestFixture,
+) -> gateway_core::RequestToolCardinality {
+    if fixture.error_code.is_some() {
+        return gateway_core::RequestToolCardinality {
+            referenced_mcp_server_count: Some(1),
+            exposed_tool_count: Some(8),
+            invoked_tool_count: Some(0),
+            filtered_tool_count: Some(2),
+        };
+    }
+
+    match fixture.component {
+        "tool-runner" => gateway_core::RequestToolCardinality {
+            referenced_mcp_server_count: Some(4),
+            exposed_tool_count: Some(27),
+            invoked_tool_count: Some(6),
+            filtered_tool_count: Some(5),
+        },
+        "customer-brief" | "prototype-review" | "design-critique" => {
+            gateway_core::RequestToolCardinality {
+                referenced_mcp_server_count: Some(3),
+                exposed_tool_count: Some(18),
+                invoked_tool_count: Some(4),
+                filtered_tool_count: Some(3),
+            }
+        }
+        "longform-eval" | "bedrock-routing" | "metric-synthesis" => {
+            gateway_core::RequestToolCardinality {
+                referenced_mcp_server_count: Some(2),
+                exposed_tool_count: Some(14),
+                invoked_tool_count: Some(2),
+                filtered_tool_count: Some(1),
+            }
+        }
+        "incident-review" | "runbook-update" => gateway_core::RequestToolCardinality {
+            referenced_mcp_server_count: Some(1),
+            exposed_tool_count: Some(9),
+            invoked_tool_count: Some(1),
+            filtered_tool_count: Some(0),
+        },
+        _ => gateway_core::RequestToolCardinality {
+            referenced_mcp_server_count: Some(0),
+            exposed_tool_count: Some(0),
+            invoked_tool_count: Some(0),
+            filtered_tool_count: Some(0),
+        },
+    }
+}
