+++ crates/gateway/src/local_demo_seed/users.rs
@@ -1,0 +1,52 @@
+use super::LocalDemoUserFixture;
+
+pub(super) const LOCAL_DEMO_USERS: &[LocalDemoUserFixture] = &[
+    LocalDemoUserFixture {
+        email: "alice@platform.local",
+    },
+    LocalDemoUserFixture {
+        email: "ben@platform.local",
+    },
+    LocalDemoUserFixture {
+        email: "cara@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "diego@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "erin@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "faye@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "gina@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "hugo@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "ivy@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "jules@research.local",
+    },
+    LocalDemoUserFixture {
+        email: "frank@applied.local",
+    },
+    LocalDemoUserFixture {
+        email: "grace@applied.local",
+    },
+    LocalDemoUserFixture {
+        email: "heidi@applied.local",
+    },
+    LocalDemoUserFixture {
+        email: "ivan@operations.local",
+    },
+    LocalDemoUserFixture {
+        email: "judy@operations.local",
+    },
+    LocalDemoUserFixture {
+        email: "mallory@operations.local",
+    },
+];
